Move A11yAutomation test-specific implementations to an A11yAutomationUtilities class in AutomationTests.
