+++ src/Desktop/UIAutomation/A11yAutomation.cs
@@ -155,35 +155,6 @@ namespace Axe.Windows.Desktop.UIAutomation
             return elements;
         }
 
-        internal static DesktopElement GetFocusedElement()
-        {
-            IUIAutomation uiAutomation = GetDefaultInstance().UIAutomation;
-            IUIAutomationElement focusedElement = uiAutomation.GetFocusedElement();
-            return new DesktopElement(focusedElement, keepElement: true, setMembers: true);
-        }
-
-        internal static DesktopElement GetDepthFirstLastLeafControlElement(DesktopElement rootElement)
-        {
-            var walker = GetDefaultInstance().GetTreeWalker(TreeViewMode.Control);
-            try
-            {
-                IUIAutomationElement leafElement = (IUIAutomationElement)rootElement.PlatformObject;
-                for (IUIAutomationElement currentElement = walker.GetLastChildElement(leafElement); currentElement != null; currentElement = walker.GetLastChildElement(leafElement))
-                {
-                    Marshal.ReleaseComObject(leafElement);
-                    leafElement = currentElement;
-                }
-
-                return leafElement is null
-                    ? rootElement
-                    : new DesktopElement(leafElement, keepElement: true, setMembers: true);
-            }
-            finally
-            {
-                Marshal.ReleaseComObject(walker);
-            }
-        }
-
         /// <summary>
         /// Get DesktopElement from UIAElement interface.
         /// </summary>
